backend/pyproject.toml
[project]
name = "mumbli-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "pyjwt[crypto]>=2.12.1",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.42.0",
    "websockets>=15.0.1",
]
